pyproject.toml
[project]
name = "fairpredictor"
version = "0.4.0"
description = "A package for running predictions using fAIr"
authors = [
    {name = "kshitijrajsharma", email = "skshitizraj@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ai-edge-litert>=1.2.0",
    "geomltoolkits>=2.0.0",
    "onnxruntime>=1.20.1",
    "opencv-python-headless>=4.10",
    "pillow>=9.1.0",
    "pydantic>=2.0",
    "requests>=2.32",
]

[project.optional-dependencies]
tensorflow = ["tensorflow>=2.15"]
pytorch = ["ultralytics>=8.0", "torch"]
all = ["fairpredictor[tensorflow,pytorch]"]

[dependency-groups]
api = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.0.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.33.0",
]
dev = [
    "commitizen>=3.31.0",
    "pre-commit>=4.0",
    "pytest>=8.3.5",
    "ruff>=0.15",
]
load-test = [
    "locust>=2.25.0",
]
notebook = [
    "contextily>=1.6.2",
    "matplotlib>=3.9.4",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]

[tool.ty.rules]
unused-ignore-comment = "ignore"
unused-type-ignore-comment = "ignore"

[tool.ty.environment]
python-version = "3.10"

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
tag_format = "v$version"
version_scheme = "semver"
update_changelog_on_bump = true
major_version_zero = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
only-include = ["predictor"]
packages = ["predictor"]

[tool.hatch.build.targets.sdist]
include = [
    "predictor/*.py",
    "predictor/yoloseg/*.py",
]
exclude = [
    "/tests",
]

[tool.uv.sources]
geomltoolkits = { path = "../geoml-toolkits", editable = true }

[build-system]
requires = ["uv-build>=0.6.6"]
build-backend = "uv_build"
